client-api: Add support for appservice /sync
While appservices are not intended to use the `/sync` endpoint, some
features such as end-to-end-encryption currently require it. Unlike
regular users however, the `/sync` endpoint takes an extra user_id
argument, which specifies the user ID of the virtual user to sync
as.[^1]

the user_id parameter to the sync endpoint is enabled with the feature
flag `appservice`, which is automatically enabled when you enable the
`helper` feature in `ruma-appservice-api`.

[^1]: https://spec.matrix.org/v1.2/application-service-api/#using-sync-and-events